pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ara"
version = "0.1.0"
description = "Ara Voice Assistant - Privacy-first, offline-capable voice assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Ammar Darkazanli", email = "adarkazanli@example.com"}
]
keywords = ["voice-assistant", "offline", "raspberry-pi", "llm", "whisper", "piper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Audio
    "pyaudio>=0.2.14",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",

    # Wake Word
    "pvporcupine>=3.0.0",

    # Speech-to-Text
    "faster-whisper>=1.0.0",

    # Language Model
    "ollama>=0.1.0",
    "httpx>=0.25.0",

    # Text-to-Speech
    "piper-tts>=1.2.0",

    # Search (optional internet features)
    "duckduckgo-search>=4.0.0",

    # Configuration
    "pyyaml>=6.0",

    # Logging & Storage
    "aiosqlite>=0.19.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.1.0",

    # Linting & Formatting
    "ruff>=0.1.0",
    "mypy>=1.5.0",

    # Type stubs
    "types-PyYAML>=6.0.0",
]

cloud = [
    # Claude API for cloud features
    "anthropic>=0.18.0",
]

[project.scripts]
ara = "ara.__main__:main"

[project.urls]
Homepage = "https://github.com/adarkazanli/RPVA"
Repository = "https://github.com/adarkazanli/RPVA"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ara*"]

[tool.setuptools.package-dir]
"" = "src"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as benchmarks",
    "integration: marks tests as integration tests",
    "requires_audio: marks tests that require audio hardware",
]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.isort]
known-first-party = ["ara"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "pyaudio.*",
    "sounddevice.*",
    "pvporcupine.*",
    "faster_whisper.*",
    "piper.*",
    "ollama.*",
    "duckduckgo_search.*",
]
ignore_missing_imports = true

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
